requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
APScheduler==3.7.0
requests==2.26.0
python-dotenv==0.19.2
gunicorn==20.1.0  # Render.comでのデプロイ用
